This PR improves `Dup` and `Swap`. 

### Remarks

When benchmarking I took a loot at `PushZero`, `PushOne` and a few others. Additionally, I took a look at IL behind `VirtualMachine.ExecuteCall` and found that it has over 250 variables. I did a sketch and removed almost 100 of them by merging variables of the same type, used in different instructions (like, `a`, `b` and `c` for both, `ADD` or `SUB`). A sample contract executing till it's out of gas showed no impact on the call. This made me undo the merging of variables, leaving the PR without them.

I added a few benchmars for other operations.

I added a `Word` struct that helped me in rewriting Dup and Swap without bloating the code with Write/Read unaligned for `ulong`s. I think this might be a possible approach for the whole `EvmStack`, as it always works on 32 bytes long words. This is a subject to a further discussion, whether or not it's worth to pursue.

### Setup

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.18363.1316 (1909/November2018Update/19H2)
Intel Core i7-6700HQ CPU 2.60GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=5.0.103
  [Host]        : .NET Core 5.0.3 (CoreCLR 5.0.321.7212, CoreFX 5.0.321.7212), X64 RyuJIT
  .NET Core 5.0 : .NET Core 5.0.3 (CoreCLR 5.0.321.7212, CoreFX 5.0.321.7212), X64 RyuJIT

Job=.NET Core 5.0  Runtime=.NET Core 5.0  

```

### Before

| Method |     Mean |    Error |   StdDev | Code Size | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |---------:|---------:|---------:|----------:|------:|------:|------:|----------:|
|   Swap | 22.53 ns | 0.327 ns | 0.290 ns |    1226 B |     - |     - |     - |         - |
|    Dup | 14.08 ns | 0.449 ns | 1.303 ns |    1229 B |     - |     - |     - |         - |


### After

| Method |     Mean |    Error |   StdDev | Code Size | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |---------:|---------:|---------:|----------:|------:|------:|------:|----------:|
|   Swap | 11.55 ns | 0.226 ns | 0.211 ns |     718 B |     - |     - |     - |         - |
|    Dup | 10.55 ns | 0.194 ns | 0.406 ns |     792 B |     - |     - |     - |         - |